# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Business Content API",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
...
        "/reviews": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Review show all or by user id or by startup id",
                "operationId": "f6cb6efd5c4ab04cb459652696b15842",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User id"
                    },
                    {
                        "name": "startup_id",
                        "in": "query",
                        "description": "Startup id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Review"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/review": {
            "put": {
                "tags": [
                    "Review"
                ],
                "summary": "Review update",
                "operationId": "a352015df05765e92bf97c51b18860f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Review"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Review"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "Create review",
                "operationId": "00e0abb5d16d3ff6397998313d8da818",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Review"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Review"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/review/{id}": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Get review by ID",
                "operationId": "d314d0653db374ce003f3f1082266c2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Review id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Review"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Review"
                ],
                "summary": "Delete review",
                "operationId": "06c8f898c04e10704c33dd482970c6ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Review id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/reward_users": {
            "get": {
                "tags": [
                    "RewardUser"
                ],
                "summary": "RewardUser show all or by user id or by reward id",
                "operationId": "790ae89994c476caa35c1bfc31817c6b",
                "parameters": [
                    {
                        "name": "reward_id",
                        "in": "query",
                        "description": "Reward id"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RewardUser"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/reward_user": {
            "put": {
                "tags": [
                    "RewardUser"
                ],
                "summary": "RewardUser update",
                "operationId": "9218359d74eb9a8fb277e346400a9266",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RewardUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RewardUser"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "RewardUser"
                ],
                "summary": "Create RewardUser",
                "operationId": "906039273d5c24103b9557d677ea80fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RewardUser"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RewardUser"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Review": {
                "title": "Review",
                "description": "Review model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "example": 1
                    },
                    "startup_id": {
                        "title": "startup_id",
                        "description": "startup_id",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "user_id",
                        "type": "integer",
                        "example": 1
                    },
                    "direction": {
                        "title": "direction",
                        "description": "direction (UTOS or STOU)",
                        "type": "string",
                        "example": "UTOS"
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "example": "Good work!"
                    },
                    "type": {
                        "title": "type",
                        "description": "type (POSITIVE or NEGATIVE)",
                        "type": "string",
                        "example": "POSITIVE"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created_at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-01-22T21:34:30.000000"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated_at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-01-22T21:34:30.000000"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Review"
                }
            },
            "RewardUser": {
                "title": "RewardUser",
                "description": "RewardUser model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "example": 1
                    },
                    "reward_id": {
                        "title": "reward_id",
                        "description": "reward_id",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "user_id",
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "title": "status",
                        "description": "status (CREATED or SENT or RECIEVED)",
                        "type": "string",
                        "example": "CREATED"
                    },
                    "destination": {
                        "title": "destination",
                        "description": "destination",
                        "type": "string",
                        "example": "Wall Street/8/16"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created_at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-01-22T21:34:30.000000"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated_at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-01-22T21:34:30.000000"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RewardUser"
                }
            }
        }
    },
...
}